move rosaria passives to asc.go and refactor rosaria a1 into AttackCB
