+++ internal/characters/rosaria/burst.go
@@ -5,9 +5,6 @@ import (
 	"github.com/genshinsim/gcsim/pkg/core/action"
 	"github.com/genshinsim/gcsim/pkg/core/attributes"
 	"github.com/genshinsim/gcsim/pkg/core/combat"
-	"github.com/genshinsim/gcsim/pkg/core/glog"
-	"github.com/genshinsim/gcsim/pkg/core/player/character"
-	"github.com/genshinsim/gcsim/pkg/modifier"
 )
 
 var burstFrames []int
@@ -97,33 +94,7 @@ func (c *char) Burst(p map[string]int) action.ActionInfo {
 		}
 	}, 56)
 
-	// Handle A4
-	// Casting Rites of Termination increases CRIT RATE of all nearby party members, excluding Rosaria herself, by 15% of Rosaria's CRIT RATE for 10s. CRIT RATE bonus gained this way cannot exceed 15%.
-	// Uses the snapshot generated by hit #1 to ensure all mods are accounted for.
-	// Confirmed via testing that mods like Rosaria A1 are accounted for, and Blizzard Strayer modifications are not
-	crit_share := 0.15 * c.Stat(attributes.CR)
-	if crit_share > 0.15 {
-		crit_share = 0.15
-	}
-
-	m := make([]float64, attributes.EndStatType)
-	m[attributes.CR] = crit_share
-	for i, char := range c.Core.Player.Chars() {
-		// skip Rosaria
-		if i == c.Index {
-			continue
-		}
-		char.AddStatMod(character.StatMod{
-			Base:         modifier.NewBaseWithHitlag("rosaria-a4", 600),
-			AffectedStat: attributes.CR,
-			Amount: func() ([]float64, bool) {
-				return m, true
-			},
-		})
-	}
-	c.Core.Log.NewEvent("Rosaria A4 activation", glog.LogCharacterEvent, c.Index).
-		Write("ends_on", c.Core.F+600).
-		Write("crit_share", crit_share)
+	c.a4()
 
 	c.SetCD(action.ActionBurst, 15*60)
 	c.ConsumeEnergy(6)
