+++ internal/characters/rosaria/skill.go
@@ -5,9 +5,6 @@ import (
 	"github.com/genshinsim/gcsim/pkg/core/action"
 	"github.com/genshinsim/gcsim/pkg/core/attributes"
 	"github.com/genshinsim/gcsim/pkg/core/combat"
-	"github.com/genshinsim/gcsim/pkg/core/glog"
-	"github.com/genshinsim/gcsim/pkg/core/player/character"
-	"github.com/genshinsim/gcsim/pkg/modifier"
 )
 
 var skillFrames []int
@@ -40,6 +37,12 @@ func (c *char) Skill(p map[string]int) action.ActionInfo {
 		HitlagFactor:       0.01,
 		CanBeDefenseHalted: false,
 	}
+	// We always assume that A1 procs on hit 1 to simplify
+	//TODO: does this need to change if we add player position?
+	var a1cb combat.AttackCBFunc
+	if p["nobehind"] != 1 {
+		a1cb = c.a1()
+	}
 	var c4cb combat.AttackCBFunc
 	if c.Base.Cons >= 4 {
 		c.c4completed = false
@@ -50,26 +53,9 @@ func (c *char) Skill(p map[string]int) action.ActionInfo {
 		combat.NewBoxHitOnTarget(c.Core.Combat.Player(), combat.Point{Y: -1}, 2, 4),
 		skillHitmark,
 		skillHitmark,
+		a1cb,
 		c4cb,
 	)
-
-	// A1 activation
-	// When Rosaria strikes an opponent from behind using Ravaging Confession, Rosaria's CRIT RATE increases by 12% for 5s.
-	// We always assume that it procs on hit 1 to simplify
-	//TODO: does this need to change if we add player position?
-	if p["nobehind"] != 1 {
-		m := make([]float64, attributes.EndStatType)
-		m[attributes.CR] = 0.12
-		c.AddStatMod(character.StatMod{
-			Base:         modifier.NewBaseWithHitlag("rosaria-a1", 300+skillHitmark),
-			AffectedStat: attributes.CR,
-			Amount: func() ([]float64, bool) {
-				return m, true
-			},
-		})
-		c.Core.Log.NewEvent("Rosaria A1 activation", glog.LogCharacterEvent, c.Index).
-			Write("ends_on", c.Core.F+300+skillHitmark)
-	}
 
 	// Rosaria E is dynamic, so requires a second snapshot
 	//TODO: check snapshot timing here
